requirements/base.txt
Django==1.9.7
django-braces==1.4.0
django-extensions==1.6.7
django-model-utils==2.5
django-reversion==2.0.3
ipython==4.2.0
logutils==0.3.3
django-pipeline==1.5.4
sphinx-rtd-theme==0.1.6
Sphinx==1.2.3
django-modeltranslation==0.11
easy-thumbnails==2.2
django-markupfield==1.3.0
markdown==2.6.1
django-rosetta==0.7.11
icalendar==3.9.1
django-suit==0.2.18